requirements.txt
Django==1.5.1
Markdown==2.3.1
Pillow==2.5.0
South==0.7.6
argparse==1.2.1
boto==2.9.2
distribute==0.6.24
django-appconf==0.6
django-debug-toolbar==0.9.4
django-imagekit==2.0.4
django-storages==1.1.8
django-taggit==0.10a1
djrill==0.5.0
geopy==0.95.1
newrelic>=1,<=2
postmonkey==1.0a4
python-linkedin>=2.0
requests==2.3.0
six==1.3.0
twitter==1.9.4
wsgiref==0.1.2
